requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
httpx>=0.25.0
python-dotenv>=1.0.0
coincurve>=20.0.0
